adding docstring, request modifications and moving the DelegateFunctionCheckpoint to checkpoiting
