+++ firedrake/adjoint/checkpointing.py
@@ -303,3 +303,20 @@ class CheckpointFunction(CheckpointBase):
 def maybe_disk_checkpoint(function):
     """Checkpoint a Function to disk if disk checkpointing is active."""
     return CheckpointFunction(function) if disk_checkpointing() else function
+
+
+class DelegatedFunctionCheckpoint(CheckpointBase):
+    """A wrapper which delegates the checkpoint of this Function to another Function.
+
+    This enables us to avoid checkpointing a Function twice when it is copied.
+
+    Parameters
+    ----------
+    other: BlockVariable
+        The block variable to which we delegate checkpointing.
+    """
+    def __init__(self, other):
+        self.other = other
+
+    def restore(self):
+        return self.other.saved_output
